requirements.txt
asgiref==3.7.2
attrs==23.2.0
Django==5.0.1
djangorestframework==3.14.0
drf-spectacular==0.27.1
drf-spectacular-sidecar==2024.1.1
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
rpds-py==0.17.1
sqlparse==0.4.4
uritemplate==4.1.1
